server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
numpy
gunicorn